Add URL ingestion job for processing video resources and update resource status

+++ KnowledgeOS.Backend/Entities/Resources/ConcreteResources/VideoResource.cs
@@ -7,5 +7,5 @@ public class VideoResource:Resource
     [MaxLength(200)]
     public string ChannelName { get; set; } = string.Empty;
     public TimeSpan? Duration { get; set; } 
-    public int ViewCount { get; set; } 
+    public long  ViewCount { get; set; } 
 }
